Fixes broken links in Client/Concept API References (#164)
## What is the goal of this PR?
The links included in Client and Concept API reference pages are now healthy and functional.

## What are the changes implemented in this PR?
As a temporary and quick solution, the links contained within the reference `.yml` files are altered to support the web platform. The change breaks the links when viewed on Github, which will be resolved once a long-term resolution is accomplished for this problem. Issued [here](https://github.com/graknlabs/docs/issues/165).