improves implementation of links_test.py (#184)
## What is the goal of this PR?
The coverage of testing links within docs is expanded to spot markdown links with an invalid format.
This means, a link found in a `.md` file that has a format other than what follows, fails the test.
- `[link](../path)`
- `[link](#anchor)`
- `[link](http://path)`
- `[link](https://path)`

## What are the changes implemented in this PR?
The regex to spot invalid link formats is checked on every `.md` file. If any invalid link is found, the test fails and the  messages are printed out to help the contributor find the invalid links.

To expand the coverage of this test further, we should also verify that the path of an internal link, always contains `.md`.