Updated new Maven artifact IDs (#192)
## What is the goal of this PR?

When deploying Maven JARs to Sonatype, the JARs are renamed to be `artifact-id-version.jar`, without the `group-id`. This is done in Sonatype because the JARs are organised into folders where the folder names represent the Group IDs. However, in certain build systems, e.g. Gradle, the JARs are collected under one directory, which means the JAR names are not unique enough to disambiguate itself from each other. We have now renamed the artifact ID of our Maven packages to be prefixed with `grakn-` to solve this problem.

## What are the changes implemented in this PR?

Prefixed artifact IDs of Maven packages with `grakn-` and `graql-`.