Fix issue with example 5 query list index in Python version (#242)
## What is the goal of this PR?

The Python client version of query 5 in the `phone_calls` example was missing the index, causing this method to fail. This is updated to be in line with the Java and Node.js  client versions of the same query.

## What are the changes implemented in this PR?

I simply referenced the zero-indexed element of the result list, same as the Java and Node.js versions.