Adding instructions to upgrade to 1.6.x (#251)
## What is the goal of this PR?

We have changed the cluster name to `Grakn Cluster` from `Grakn Example Cluster` in `cassandra.yaml`. The cluster name is persisted in the data and would cause an issue if you're upgrading an existing installation.

Therefore, this PR provides the instruction for people to update the cluster name of an existing data so that they can upgrade to Grakn Core 1.6.x.