Workbase explanation warning (#257)
## What is the goal of this PR?
In recent Grakn core versions 1.6.0 and up, explanations have been changed to be fetched (https://github.com/graknlabs/grakn/pull/5483). A corresponding change in Workbase meant that some relations are not included in the _explain_ functionality of workbase. To work around this, for now, we leave a note in the Workbase docs.

## What are the changes implemented in this PR?
* Add workbase Important Note