Repair Hanging Tests (#489)
## What is the goal of this PR?

A number of tests were hanging and/or failing due to failing to clear up a schema session after a failing step. We made certain that the session closes in these cases, and fixed a number of bugs that had been disguised by this.

## What are the changes implemented in this PR?

Wrapped Language tests in a large try finally that closes session, fixed a number of broken examples.