Explain best practice when providing addresses to TypeDB Cluster client (#575)
## What is the goal of this PR?

The docs stated that the user should supply all server addresses when instantiating TypeDB Cluster client, which is not technically correct. We clarify the behaviour and describe best practice.

## What are the changes implemented in this PR?

We updated the description to state that supplying one server address is usually sufficient, while supplying all addresses is best practice.