Indicate that getPlayers takes in a list of roles (#579)
## What is the goal of this PR?

`Relation.getPlayers` was marked as taking in a single `RoleType` in its parameters. It is now correctly marked as taking in a List (or Array) of `RoleType`s.

## What are the changes implemented in this PR?

`Relation.getPlayers` was marked as taking in a single `RoleType` in its parameters. It is now correctly marked as taking in a List (or Array) of `RoleType`s; fixes #576