Update outputs in usage of python examples' template functions (#597)
## What is the goal of this PR?

In the "Migrate - Python" section under Examples, the results shown for the template functions did not match the output that the functions returned. The update attempts to match the expected output to the output stated in the docs. In addition, one of the json tabs had a reference to the `xml` python library, so the reference was moved to the xml tab. Also the section headers for the template functions were in camelCase and didn't match the python section's snake_case format.

## What are the changes implemented in this PR?

- In `08-examples/04-phone-calls-migration-python.md`, the outputs of the `person_template` function have been updated to match what the function produces
- The reference to the `xml` package in python was moved from the json tab to the xml tab
- The section headers for the template functions was reformatted from camelCase to snake_case to match the function names in python