Fix Typo
## What is the goal of this PR?

We fix a typo in the link that was overlooked in https://github.com/vaticle/docs/pull/610

## What are the changes implemented in this PR?

The typo is fixed, the link points to the correct destination. Furthermore, we have opened an issue to look at preventing this class of errors in the future: https://github.com/vaticle/docs/issues/612.